# base-go/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {
            "name": "Direktorat Pengembangan Teknologi dan Sistem Informasi (DPTSI) - ITS",
            "url": "http://its.ac.id/dptsi",
            "email": "dptsi@its.ac.id"
        }
    },
    "paths": {
...
        "/sdm": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SDM"
                ],
                "summary": "List semua SDM",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.SDM"
                            }
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SDM"
                ],
                "summary": "Tambah SDM",
                "parameters": [
                    {
                        "description": "Data SDM",
                        "name": "sdm",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.SDM"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.SDM"
                        }
                    }
                }
            }
        },
        "/sdm/{nik}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SDM"
                ],
                "summary": "Get SDM by NIK",
                "parameters": [
                    {
                        "type": "string",
                        "description": "NIK",
                        "name": "nik",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.SDM"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SDM"
                ],
                "summary": "Update SDM by NIK",
                "parameters": [
                    {
                        "type": "string",
                        "description": "NIK",
                        "name": "nik",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Data SDM",
                        "name": "sdm",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.SDM"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.SDM"
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SDM"
                ],
                "summary": "Hapus SDM by NIK",
                "parameters": [
                    {
                        "type": "string",
                        "description": "NIK",
                        "name": "nik",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.SDM": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "enabled": {
                    "type": "boolean"
                },
                "id_sdm": {
                    "type": "string"
                },
                "nama": {
                    "type": "string"
                },
                "nik": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "CSRF Token": {
            "description": "CSRF token yang didapatkan dari browser -\u003e inspect element -\u003e application -\u003e storage -\u003e cookies -\u003e CSRF-TOKEN (Untuk firefox, storage berada pada tab tersendiri)",
            "type": "apiKey",
            "name": "x-csrf-token",
            "in": "header"
        }
    },
...
}